refactor: move data analysis files into tools/analysis/ subdirectory
