+++ source/tests/unit/tools/test_tools_hypothesistester.cpp
@@ -2,7 +2,7 @@
 
 #include <cmath>
 
-#include "tools/HypothesisTesterDefaultImpl1.h"
+#include "tools/analysis/HypothesisTesterDefaultImpl1.h"
 
 namespace {
