+++ source/tools/analysis/CMakeLists.txt
@@ -1,0 +1,2 @@
+# Analysis tool sources will be added here.
+# This subdirectory is included from source/tools/CMakeLists.txt.
