+++ source/tools/analysis/DataAnalyserDefaultImpl.cpp
