+++ source/tools/analysis/DataAnalyserDefaultImpl.h
