+++ source/tools/analysis/FitterDefaultImpl.h
@@ -3,8 +3,8 @@
 
 #include "DatasetLoader.h"
 #include "Fitter_if.h"
-#include "ProbabilityDistributionBase.h"
-#include "SolverDefaultImpl1.h"
+#include "tools/ProbabilityDistributionBase.h"
+#include "tools/SolverDefaultImpl1.h"
 
 #include <algorithm>
 #include <cmath>
