+++ source/tools/analysis/HypothesisTesterDefaultImpl1.cpp
@@ -11,8 +11,8 @@
  */
 
 #include "HypothesisTesterDefaultImpl1.h"
-#include "ProbabilityDistribution.h"
-#include "SolverDefaultImpl1.h"
+#include "tools/ProbabilityDistribution.h"
+#include "tools/SolverDefaultImpl1.h"
 #include "kernel/statistics/StatisticsDataFile_if.h"
 #include "kernel/TraitsKernel.h"
 #include <algorithm>
